Fix SC2115 warning to avoid unexpected deletion of system's root directory
Warning:

  rm -rf "${app_dir}"/*
         ^------------^ SC2115 (warning): Use "${var:?}" to ensure this never expands to /* .

See https://www.shellcheck.net/wiki/SC2115

+++ sign.bash
@@ -211,7 +211,7 @@ then
 fi
 
 log "Copy signed files back to ${app_dir}"
-rm -rf "${app_dir}"/*
+rm -rf "${app_dir:?}"/*
 cp -aR "${tmp_app_dir}"/* "${app_dir}"/
 
 log "Generating PKG"
